Full training v12: adapter architecture validated through overfitting + 10K step full dataset training
- 5-layer adapter injection [20,22,24,26,28] with tanh gating (init=0.5, frozen)
- Unfreeze last 5 backbone blocks (backbone_lr=5e-6) for adapter adaptation
- Multi-CFG dropout (10% drop each context) for classifier-free guidance
- 8-bit AdamW optimizer, uniform flow matching time sampling
- Overfitting validation: 24.71 dB PSNR on single sample
- Full dataset (1982 sequences): trained to step 10000
- Add overfit test configs and verification scripts

+++ scripts/run_ablation_video.py
@@ -154,10 +154,12 @@ def generate_shot_single_cfg(
 
             v_out = v_null + guidance_scale * (v_cond - v_null)
 
+            # Cast to float32 for numerical precision in scheduler step
+            v_out = v_out.float()
             step_out = pipeline.diffusion.inference_step(
-                v_out, x, t, i, timesteps, state=state,
+                v_out, x.float(), t, i, timesteps, state=state,
             )
-            x = step_out.latents
+            x = step_out.latents.to(torch.bfloat16)
             state = step_out.state
 
         del text_embeds, null_text_embeds, ctx_cond, mask_cond
@@ -230,10 +232,12 @@ def generate_shot_multi_cfg(
                      + omega_local  * (v_local - v_text)
                      + omega_global * (v_glob  - v_text))
 
+            # Cast to float32 for numerical precision in scheduler step
+            v_out = v_out.float()
             step_out = pipeline.diffusion.inference_step(
-                v_out, x, t, i, timesteps, state=state,
+                v_out, x.float(), t, i, timesteps, state=state,
             )
-            x = step_out.latents
+            x = step_out.latents.to(torch.bfloat16)
             state = step_out.state
 
         del text_embeds, null_text_embeds
